Allow to use a self-signed certificate as initial certificate
When initialy creates a certificate, there is a chicken/egg problem.
The webserver refuses to start without a valid certificate, but acme_tiny needs
a webserver to generate a certificate.
For such cases the it might be a good idea to create a dummy certificate which is self-signed
and only valid for one day.
It will be replaced by a real valid certificate when renew.py runs for the first time.